requirements.txt
prettytable
pymupdf
python-docx